Update warning fixes based on review comments
- Remove SMALL_FOOTPRINT code from SSE 4.1 FIR implementation, matching
  the C implementation.
- Fix GCC -O3 false positive: silk/NLSF_del_dec_quant.c:137:20:
  warning: array subscript is above array bounds [-Warray-bounds]

+++ src/opus_compare.c
@@ -171,7 +171,7 @@ int main(int _argc,const char **_argv){
   float   *X;
   float   *Y;
   double    err;
-  double   Q;
+  float    Q;
   size_t   xlength;
   size_t   ylength;
   size_t   nframes;
@@ -346,7 +346,7 @@ int main(int _argc,const char **_argv){
           float re;
           float im;
           re=Y[(xi*yfreqs+xj)*nchannels+ci]/X[(xi*NFREQS+xj)*nchannels+ci];
-          im=re-(float)log(re)-1;
+          im=(float)(re-log(re)-1);
           /*Make comparison less sensitive around the SILK/CELT cross-over to
             allow for mode freedom in the filters.*/
           if(xj>=79&&xj<=81)im*=0.1F;
@@ -364,7 +364,7 @@ int main(int _argc,const char **_argv){
     err+=Ef*Ef;
   }
   err=pow(err/nframes,1.0/16);
-  Q=100*(1-0.5*log(1+err)/log(1.13));
+  Q=(float)(100*(1-0.5*log(1+err)/log(1.13)));
   if(Q<0){
     fprintf(stderr,"Test vector FAILS\n");
     fprintf(stderr,"Internal weighted error is %f\n",err);
